move userContext to context folder and update imports
