+++ app/_layout.tsx
@@ -1,6 +1,6 @@
 import { GestureHandlerRootView } from "react-native-gesture-handler";
 import { Stack } from "expo-router";
-import { UserProvider } from "@/utils/UserContext";
+import { UserProvider } from "@/contexts/UserContext";
 
 export default function RootLayout() {
   return (
